backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
motor==3.7.1
pymongo==4.10.1
python-dotenv==1.0.1
python-multipart==0.0.17
pyjwt==2.10.1
requests==2.32.3
aiohttp==3.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==24.1.0
PyPDF2==3.0.1
python-docx==1.1.0